cleanup: moved files around and removed unsued functions/components
