+++ WebApp/src/components/header.jsx
@@ -9,7 +9,6 @@
  ***************************************************************/
 
 import { useState, useEffect } from "preact/hooks";
-import { api, HTTP_STATUS } from "../utils/utils";
 import { Match } from "preact-router/match";
 import logo from "../images/cymaspaceLogo.jpg"
 
@@ -18,7 +17,6 @@ import logo from "../images/cymaspaceLogo.jpg"
  *
  * @returns Header component
  */
-
 const Header = ({ setCurrentPage }) => {
   const NavLink = ({ to, children }) => {
 
@@ -44,20 +42,12 @@ const Header = ({ setCurrentPage }) => {
     );
   };
 
-  const printMemory = async () => {
-    const res = await api("GET", "/dev/getMemory");
-
-    if (res?.status == HTTP_STATUS.OK) {
-      console.log("Check serial monitor");
-    }
-  };
-
   return (
     <div className="p-4 flex w-full min-h-[8vh] justify-between items-center">
       
       {/* Left side */}
       <div className="flex flex-row justify-between items-center">
-        <img src={logo} alt="logo" className="w-10 h-10"></img> {/** TODO: add actual alt text */}
+        <img src={logo} alt="Image of the Cymaspace logo." className="w-10 h-10"></img> {/** TODO: add actual alt text */}
         <h2 className="font-bold text-2xl ml-2">Vibrosonics</h2>
       </div>
 
@@ -73,14 +63,6 @@ const Header = ({ setCurrentPage }) => {
             <NavLink to="https://www.youtube.com/@cymaspace">CymaSpace</NavLink>
             <NavLink to="https://github.com/udellc/Vibrosonics">GitHub Repo</NavLink>
           </nav>
-
-          {/** remove? 
-          <div className="pt-4">
-            DEBUGGING
-            <button className="ml-3 bg-cyan-400 cursor-pointer" onClick={printMemory}>
-              PRINT MEMORY BTN
-            </button>
-          </div>*/}
         </div>
       </div>
     </div>
