+++ main/webInterface.h
@@ -1,5 +1,5 @@
 /***************************************************************
- * FILE: webServer.h
+ * FILE: webInterface.h
  * 
  * DATE: 11/18/2025
  *
